requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
openai==1.12.0
pytest==8.0.2
httpx==0.27.0 